+++ test/services/build_search_query_test.exs
@@ -16,7 +16,7 @@ defmodule Rummage.Ecto.Services.BuildSearchQueryTest do
   @search_types ~w{like ilike eq gt lt gteq lteq in}a
   @search_exprs ~w{where or_where not_where}a
 
-  test "Single Interpolation Fragments" do
+  test "Definitions of Single Interpolation Fragments" do
     for fragment <- @supported_fragments_one do
       for search_type <- @search_types do
         for search_expr <- @search_exprs do
@@ -43,7 +43,30 @@ defmodule Rummage.Ecto.Services.BuildSearchQueryTest do
               [queryable, {:fragment, fragment, :field}, term, search_expr])
       end
 
-      search_type = :in
+      for search_expr <- @search_exprs do
+        name = :"handle_#{search_type}"
+        queryable = Rummage.Ecto.Product
+        term = false
+        result = apply(BuildSearchQuery, name,
+              [queryable, {:fragment, fragment, :field}, term, search_expr])
+
+        assert result == apply(BuildSearchQuery, name,
+              [queryable, {:fragment, fragment, :field}, term, search_expr])
+
+      end
+
+      search_type = :is_nil
+      for search_expr <- @search_exprs do
+        name = :"handle_#{search_type}"
+        queryable = Rummage.Ecto.Product
+        term = true
+        result = apply(BuildSearchQuery, name,
+              [queryable, {:fragment, fragment, :field}, term, search_expr])
+
+        assert result == apply(BuildSearchQuery, name,
+              [queryable, {:fragment, fragment, :field}, term, search_expr])
+      end
+
       for search_expr <- @search_exprs do
         name = :"handle_#{search_type}"
         queryable = Rummage.Ecto.Product
@@ -57,7 +80,7 @@ defmodule Rummage.Ecto.Services.BuildSearchQueryTest do
     end
   end
 
-  test "Double Interpolation Fragments" do
+  test "Definitions of Double Interpolation Fragments" do
     for fragment <- @supported_fragments_two do
       for search_type <- @search_types do
         for search_expr <- @search_exprs do
@@ -84,7 +107,29 @@ defmodule Rummage.Ecto.Services.BuildSearchQueryTest do
               [queryable, {:fragment, fragment, :field1, :field2}, term, search_expr])
       end
 
-      search_type = :in
+      for search_expr <- @search_exprs do
+        name = :"handle_#{search_type}"
+        queryable = Rummage.Ecto.Product
+        term = false
+        result = apply(BuildSearchQuery, name,
+              [queryable, {:fragment, fragment, :field1, :field2}, term, search_expr])
+
+        assert result == apply(BuildSearchQuery, name,
+              [queryable, {:fragment, fragment, :field1, :field2}, term, search_expr])
+      end
+
+      search_type = :is_nil
+      for search_expr <- @search_exprs do
+        name = :"handle_#{search_type}"
+        queryable = Rummage.Ecto.Product
+        term = true
+        result = apply(BuildSearchQuery, name,
+              [queryable, {:fragment, fragment, :field1, :field2}, term, search_expr])
+
+        assert result == apply(BuildSearchQuery, name,
+              [queryable, {:fragment, fragment, :field1, :field2}, term, search_expr])
+      end
+
       for search_expr <- @search_exprs do
         name = :"handle_#{search_type}"
         queryable = Rummage.Ecto.Product